changed so gui builds by default, use tag headless to not build with gui
